+++ CHANGELOG.md
@@ -6,7 +6,15 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [1.0.3] - 2024-11-24
+## [1.0.5] - 2024-11-25
+
+### Fixed
+- MonoNode register comparer issue, always node exists or no error on exists.
+
+### Add
+- add Enable/Disable Auto Register button in MonoNode. 
+
+## [1.0.4] - 2024-11-24
 
 First public release. If you have an older version, please update or re-install.   
 For detail please visit and bookmark our [GitBook](https://aceland-workshop.gitbook.io/aceland-unity-packages/)